+++ tests/Widgets/Button.cpp
@@ -67,6 +67,15 @@ TEST_CASE("[Button]")
         REQUIRE(button->getTextSize() == 25);
     }
 
+    SECTION("IgnoreKeyEvents")
+    {
+        REQUIRE(!button->getIgnoreKeyEvents());
+        button->setIgnoreKeyEvents(true);
+        REQUIRE(button->getIgnoreKeyEvents());
+        button->setIgnoreKeyEvents(false);
+        REQUIRE(!button->getIgnoreKeyEvents());
+    }
+
     SECTION("Events / Signals")
     {
         SECTION("ClickableWidget")
@@ -87,11 +96,8 @@ TEST_CASE("[Button]")
                 button->leftMouseReleased({115, 80});
                 REQUIRE(pressedCount == 0);
 
-                SECTION("mouse press")
-                {
-                    button->leftMousePressed({115, 80});
-                    REQUIRE(pressedCount == 0);
-                }
+                button->leftMousePressed({115, 80});
+                REQUIRE(pressedCount == 0);
 
                 button->leftMouseReleased({115, 80});
                 REQUIRE(pressedCount == 1);
@@ -112,6 +118,19 @@ TEST_CASE("[Button]")
                 keyEvent.code = tgui::Event::KeyboardKey::Enter;
                 button->keyPressed(keyEvent);
                 REQUIRE(pressedCount == 2);
+
+                SECTION("ignored")
+                {
+                    button->setIgnoreKeyEvents(true);
+
+                    keyEvent.code = tgui::Event::KeyboardKey::Space;
+                    button->keyPressed(keyEvent);
+                    REQUIRE(pressedCount == 2);
+
+                    keyEvent.code = tgui::Event::KeyboardKey::Enter;
+                    button->keyPressed(keyEvent);
+                    REQUIRE(pressedCount == 2);
+                }
             }
         }
     }